refactor: move operations capability into native module (#371)
