+++ internal/operations/telemetry_query_errors.go
@@ -1,4 +1,4 @@
-package service
+package operations
 
 import (
 	"context"
