+++ internal/operations/telemetry_query_errors_test.go
@@ -1,4 +1,4 @@
-package service
+package operations
 
 import (
 	"context"
