+++ internal/operations/telemetry_service_unconfigured_test.go
@@ -1,4 +1,4 @@
-package service
+package operations
 
 import (
 	"context"
