+++ internal/operations/usage_metrics_service_test.go
@@ -1,4 +1,4 @@
-package service
+package operations
 
 import (
 	"context"
